endPoints

+++ src/main/java/com/example/vantasOtm/controller/MedioPagoController.java
@@ -5,6 +5,7 @@ import com.example.vantasOtm.model.MedioPago;
 import com.example.vantasOtm.repository.MedioPagoRepository;
 import com.example.vantasOtm.service.MedioPagoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -22,21 +23,24 @@ public class MedioPagoController {
     private MedioPagoRepository medioPagoRepository;
 
 
-
     @PostMapping("/save")
         public MedioPago add(@RequestBody MedioPago medioPago){
+
+        //comprobar si se envian los datos necesarios
         if(medioPago.getMedioPago().equals("")){
-            throw new RequestException("P-401","MedioPago is required");
+            throw new RequestException("P-401", HttpStatus.INTERNAL_SERVER_ERROR,"Datos faltantes");
         }
+
         return medioPagoService.saveMedioPago(medioPago);
     }
 
     @GetMapping("/{id}")
-        public MedioPago getMedioPago(@PathVariable Integer id ){
+    public MedioPago getMedioPago(@PathVariable Integer id ){
 
-            if(!medioPagoRepository.existsById(id)){
-                throw new RequestException("P-401","Medio no existe");
-            }
+        //comprobar si existe la entidad
+        if(!medioPagoRepository.existsById(id)){
+            throw new RequestException("P-401", HttpStatus.INTERNAL_SERVER_ERROR,"Entidad no existe");
+        }
 
         return medioPagoService.getMedioPago(id);
     }
@@ -48,12 +52,29 @@ public class MedioPagoController {
 
     @DeleteMapping("/delete/{id}")
     public String deleteMedioPago(@PathVariable Integer id ){
-            medioPagoService.deleteMedioPago(id);
-            return "MedioPago eliminado";
+
+        //comprobar si existe la entidad
+        if(!medioPagoRepository.existsById(id)){
+            throw new RequestException("P-401", HttpStatus.INTERNAL_SERVER_ERROR,"Entidad no existe");
         }
+
+        medioPagoService.deleteMedioPago(id);
+        return "MedioPago eliminado";
+    }
 
     @PutMapping("/edit")
     public MedioPago deleteMedioPago(@RequestBody MedioPago medioPago ){
+
+        //comprobar si existe la entidad
+        if(!medioPagoRepository.existsById(medioPago.getId())){
+            throw new RequestException("P-401", HttpStatus.INTERNAL_SERVER_ERROR,"Entidad no existe");
+        }
+
+        //comprobar si se envian los datos necesarios
+        if(medioPago.getMedioPago().equals("")){
+            throw new RequestException("P-401", HttpStatus.INTERNAL_SERVER_ERROR,"Datos faltantes");
+        }
+
         medioPagoService.saveMedioPago(medioPago);
         return medioPago;
     }
